requirements.txt
faster-whisper>=1.1.0
pyannote.audio>=3.3
torch>=2.0
